\name{convert.vector.to.list}
\alias{convert.vector.to.list}
\title{Helper function for the graphing interface}
\usage{
convert.vector.to.list(Parameters, name = "Mean")
}
\arguments{
  \item{Parameters}{a vector of parameters}

  \item{name}{Name of the parameter}
}
\value{
a list, where each element represents an alternative and
has a Mean value
}
\description{
As named, this function takes a vector where each element
is a mean, then returns back a list, with each list item
having the mean
}

