\docType{data}
\name{Data.Election}
\alias{Data.Election}
\title{A6 Election Data}
\description{
  This is a public election dataset collected by Nicolaus
  Tideman where the voters provided partial orders on
  candidates. A partial order includes comparisons among a
  subset of alternative, and the non-mentioned alternatives
  in the partial order are considered to be ranked lower
  than the lowest ranked alternative among mentioned
  alternatives.
}
\author{
  Nicolaus Tideman
}
\keyword{datasets}

