\docType{data}
\name{Data.Election9}
\alias{Data.Election9}
\title{A9 Election Data}
\usage{
data(Data.Election9)
}
\description{
This is a public election dataset collected by Nicolaus
Tideman where the voters provided partial orders on
candidates. A partial order includes comparisons among a
subset of alternative, and the non-mentioned alternatives
in the partial order are considered to be ranked lower than
the lowest ranked alternative among mentioned alternatives.
}
\author{
Nicolaus Tideman
}

