\name{Generate.Zemel.Parameters}
\alias{Generate.Zemel.Parameters}
\title{Generates possible scores for a Zemel model}
\usage{
Generate.Zemel.Parameters(m)
}
\arguments{
  \item{m}{Number of alternatives}
}
\value{
a set of scores, all whose logs sum to 1
}
\description{
Generates possible scores for a Zemel model
}
\examples{
Generate.Zemel.Parameters(10)
}

