\name{Normal.Pairwise.Prob}
\alias{Normal.Pairwise.Prob}
\title{Pairwise Probability for Normal Model}
\usage{
Normal.Pairwise.Prob(a, b)
}
\arguments{
  \item{a}{list containing parameters for a}

  \item{b}{list containing parameters for b}
}
\value{
probability that a beats b
}
\description{
Given alternatives a and b (both items from the inference
object) what is the probability that a beats b?
}

