\name{PdfModel}
\alias{PdfModel}
\title{The PDF of a set of independent normal random variables}
\usage{
  PdfModel(X, parameter)
}
\arguments{
  \item{X}{vector of scalar values}

  \item{parameter}{list containing vetor of means for
  normal distributions, variances are set to one}
}
\value{
  value of the PDF for vector X given the parameter
}
\description{
  The PDF of a set of independent normal random variables
}

